// frontend/src/api/openapi/models/JobStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum JobStatus {
  QUEUED = "queued",
  FINISHED = "finished",
  FAILED = "failed",
  STARTED = "started",
  DEFERRED = "deferred",
  SCHEDULED = "scheduled",
  STOPPED = "stopped",
  CANCELED = "canceled",
}